Use of fix sized integer to remove cppcheck warnings

+++ ElementsKernel/src/Lib/Real.cpp
@@ -8,6 +8,7 @@
 
 #include <cassert>
 #include <cstdlib>
+#include <cstdint>
 #include "ElementsKernel/Real.h"
 
 
@@ -20,17 +21,17 @@ bool almostEqual2sComplement(const float& a, const float& b, const int& max_ulps
 {
 //    int a_int = *(int*)&a;
     // cppcheck-suppress invalidPointerCast
-    int a_int = *reinterpret_cast<const int *>(&a);
+    int32_t a_int = *reinterpret_cast<const int32_t *>(&a);
     // Make a_int lexicographically ordered as a twos-complement int
     if (a_int < 0)
         a_int = 0x80000000 - a_int;
     // Make b_int lexicographically ordered as a twos-complement int
     //    int b_int = *(int*)&b;
     // cppcheck-suppress invalidPointerCast
-    int b_int = *reinterpret_cast<const int *>(&b);
+    int32_t b_int = *reinterpret_cast<const int32_t *>(&b);
     if (b_int < 0)
         b_int = 0x80000000 - b_int;
-    int int_diff = ::abs(a_int - b_int);
+    int32_t int_diff = ::abs(a_int - b_int);
     if (int_diff <= max_ulps)
         return true;
     return false;
@@ -41,7 +42,7 @@ bool almostEqual2sComplement(const double& a, const double& b, const int& max_ul
 {
 //    long long a_int = *(long long*)&a;
     // cppcheck-suppress invalidPointerCast
-    long long a_int = *reinterpret_cast<const long long *>(&a);
+    int64_t a_int = *reinterpret_cast<const int64_t *>(&a);
     // Make a_int lexicographically ordered as a twos-complement int
     if (a_int < 0)
         a_int = 0x8000000000000000 - a_int;
@@ -49,11 +50,11 @@ bool almostEqual2sComplement(const double& a, const double& b, const int& max_ul
     // Make b_int lexicographically ordered as a twos-complement int
 //    long long b_int = *(long long*)&b;
     // cppcheck-suppress invalidPointerCast
-    long long b_int = *reinterpret_cast<const long long*>(&b);
+    int64_t b_int = *reinterpret_cast<const int64_t*>(&b);
     if (b_int < 0)
         b_int = 0x8000000000000000 - b_int;
 
-    long long int_diff = ::abs(a_int - b_int);
+    int64_t int_diff = ::abs(a_int - b_int);
     if (int_diff <= max_ulps && -max_ulps <= int_diff)
     //if (int_diff <= maxUlps)
         return true;
